CLN: rename files in c++ side
